+++ webwithpy/exeptions/HttpExceptions.py
@@ -2,6 +2,7 @@ class HttpException(Exception):
     def __init__(self, error_type: str, message: str):
         self.message = message
         self.error_type = error_type
+        super().__init__(self.__str__())
 
     def __repr__(self):
         return self.__str__()
