--- webwithpy/orm2/__init__.py
@@ -1,1 +1,0 @@
-from .objects.query import Query
