--- webwithpy/orm2/objects/table.py
@@ -1,14 +1,0 @@
-class Table:
-    pass
-
-
-class DefaultField:
-    def __init__(self): ...
-
-
-class ReferencedField(DefaultField):
-    pass
-
-
-class Field(DefaultField):
-    pass
